requirements(all).txt
defusedxml== 0.6.0
diff-match-patch== 20181111
dill== 0.2.9
distributed==2.10.0
Django==  3.0.3
docopt==  0.6.2
docutils==0.16
en-core-web-sm==2.0.0
entrypoints==0.3
et-xmlfile== 1.0.1
fast-histogram==0.7
fastcache==  1.1.0
filelock==3.0.12
flake8==  3.7.9
Flask==1.1.1
Flask-Cors== 3.0.8
Flask-Script==  2.0.6
fsspec==  0.6.2
future==  0.18.2
gast== 0.3.2
gevent==  1.4.0
glue-core==  0.15.6
glue-vispy-viewers==  0.12.2
glueviz== 0.15.2
GoogleTran== 1.0
googletrans==2.4.0
greenlet==0.4.15
grpcio==  1.16.1
h5py== 2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hypothesis== 4.54.2
identify==1.4.11
idna== 2.8
imageio== 2.6.1
imagesize==  1.2.0
importlib-metadata==  1.5.0
importlib-resources== 1.0.2
inline==  0.0.1
intervaltree==  3.0.2
ipykernel==  5.1.4
ipython== 7.12.0
ipython-genutils== 0.2.0
ipywidgets== 7.5.1
isort==4.3.21
itsdangerous==  1.1.0
jdcal==1.4.1
jedi== 0.16.0
jieba==0.42.1
Jinja2==  2.10.3
joblib==  0.14.1
json5==0.9.1
jsonify== 0.5
jsonschema== 3.2.0
jupyter== 1.0.0
jupyter-client==5.3.4
jupyter-console==  6.1.0
jupyter-core==  4.6.1
jupyterlab== 1.2.6
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==  1.0.8
Keras-Preprocessing== 1.1.0
keyring== 21.1.0
kiwisolver== 1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==  0.2.0
lxml== 4.5.0
Markdown==3.1.1
MarkupSafe== 1.1.1
mathparse==  0.1.2
matplotlib== 3.1.3
mccabe==  0.6.1
menuinst==1.4.16
mistune== 0.8.4
mkl-fft== 1.0.15
mkl-random== 1.1.0
mkl-service==2.3.0
mock== 4.0.1
more-itertools==8.0.2
mpl-scatter-density== 0.6
mpmath==  1.1.0
msgpack== 0.5.6
msgpack-numpy== 0.4.4.3
multipledispatch== 0.6.0
murmurhash== 1.0.2
nbconvert==  5.6.1
nbformat==5.0.4
networkx==2.4
nltk== 3.4.5
nodeenv== 1.3.4
nose== 1.3.7
notebook==6.0.3
numba==0.48.0
numexpr== 2.7.1
numpy==1.18.2
numpydoc==0.9.2
oauthlib==3.1.0
olefile== 0.46
opencv-python== 4.2.0.32
openpyxl==3.0.3
packaging==  20.1
pandas==  1.0.0
pandocfilters== 1.4.2
paramiko==2.7.1
parso==0.6.1
partd==1.1.0
path== 13.1.0
pathlib== 1.0.1
pathlib2==2.3.5
pathtools==  0.1.2
patsy==0.5.1
pep8== 1.7.1
pexpect== 4.8.0
pickleshare==0.7.5
Pillow==  5.2.0
Pint== 0.10.1
pip==  20.0.2
pkginfo== 1.5.0.1
plac== 0.9.6
plotly==  4.4.1
pluggy==  0.13.1
ply==  3.11
pre-commit== 2.0.1
preshed== 2.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.2
psutil==  5.6.7
py==1.8.1
pycodestyle==2.5.0
pycosat== 0.6.3
pycparser==  2.19
pycrypto==2.6.1
pycurl==  7.43.0.5
pydocstyle== 4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==  2.4.4
pymongo== 3.10.1
PyNaCl==  1.3.0
pyodbc==  4.0.0-unsupported
PyOpenGL==3.1.1a1
pyOpenSSL==  19.1.0
pyparsing==  2.4.6
pyreadline== 2.1
pyrsistent== 0.15.7
PySocks== 1.7.1
pytest==  5.3.4
pytest-arraydiff== 0.3
pytest-astropy==0.7.0
pytest-astropy-header==  0.1.1
pytest-doctestplus==  0.5.0
pytest-openfiles== 0.4.0
pytest-remotedata==0.3.2
python-dateutil==  2.7.5
python-dotenv== 0.12.0
python-jsonrpc-server==  0.3.4
python-language-server== 0.31.7
pytz== 2019.3
PyWavelets== 1.1.1
pywin32== 227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==  3.13
pyzmq==17.0.0
QDarkStyle== 2.8
QtAwesome==  0.6.1
qtconsole==  4.6.0
QtPy== 1.9.0
readme-renderer==  24.0
regex==2018.6.21
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rope== 0.16.0
Rtree==0.9.3
ruamel-yaml==0.15.87
ruamel.yaml==0.15.0
scikit-image==  0.16.2
scikit-learn==  0.22.1
scipy==1.3.2
seaborn== 0.10.0
Send2Trash== 1.5.0
setuptools== 45.1.0.post20200127
simplegeneric== 0.8.1
singledispatch==3.4.0.3
six==  1.14.0
snowballstemmer==  2.0.0
sortedcollections==1.1.2
sortedcontainers== 2.1.0
soupsieve==  1.9.5
spacy==2.1.9
Sphinx==  2.4.0
sphinx-rtd-theme== 0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==  1.0.1
sphinxcontrib-htmlhelp== 1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==  1.2.0
spyder==  4.0.1
spyder-kernels==1.8.1
SQLAlchemy== 1.3.13
sqlparse==0.3.0
srsly==1.0.1
statsmodels==0.11.0
sympy==1.5.1
tables==  3.6.1
tblib==1.6.0
tensorboard==1.14.0
tensorflow== 1.14.0
tensorflow-estimator==1.14.0
termcolor==  1.1.0
terminado==  0.8.3
testpath==0.4.4
tflearn== 0.3.2
thinc==7.0.8
toml== 0.10.0
toolz==0.10.0
tornado== 6.0.3
tox==  3.14.3
tqdm== 4.42.1
traitlets==  4.3.3
translate==  3.5.0
translator== 0.0.9
twine==3.1.1
twython== 3.7.0
typed-ast==  1.4.1
ujson==1.35
unicodecsv== 0.14.1
urllib3== 1.25.8
virtualenv== 16.7.9
waitress==1.4.3
wasabi==  0.6.0
watchdog==0.10.2
wcwidth== 0.1.8
webencodings==  0.5.1
Werkzeug==0.16.0
wheel==0.33.6
widgetsnbextension==  3.5.1
win-inet-pton== 1.1.0
win-unicode-console== 0.5
wincertstore==  0.2
wrapt==1.11.2
xlrd== 1.2.0
XlsxWriter== 1.2.7
xlwings== 0.17.1
xlwt== 1.3.0
yapf== 0.28.0
zict== 1.0.0
zipp== 2.1.0